Move _check_extra_js_files to the general linter (#11864)
* Move extra js file check

* Fix check logic

* Addressed review comments
