+++ scripts/linters/general_purpose_linter_test.py
@@ -51,6 +51,7 @@ INVALID_CSS_FILEPATH = os.path.join(LINTER_TESTS_DIR, 'invalid.css')
 # Js and Ts filepaths.
 FILE_IN_EXCLUDED_PATH = os.path.join(
     'core', 'tests', 'build_sources', 'assets', 'constants.js')
+EXTRA_JS_FILEPATH = os.path.join('core', 'templates', 'demo.js')
 INVALID_EXPLORE_FILEPATH = os.path.join(LINTER_TESTS_DIR, 'invalid_explore.js')
 INVALID_PAUSE_FILEPATH = os.path.join(LINTER_TESTS_DIR, 'invalid_pause.js')
 INVALID_SLEEP_FILEPATH = os.path.join(LINTER_TESTS_DIR, 'invalid_sleep.js')
@@ -608,6 +609,18 @@ class GeneralLintTests(test_utils.LinterTestBase):
             ['There should be a single newline at the end of file.'],
             lint_task_report.trimmed_messages)
         self.assertEqual('Newline at EOF', lint_task_report.name)
+        self.assertTrue(lint_task_report.failed)
+
+    def test_check_extra_js_file_found(self):
+        linter = general_purpose_linter.GeneralPurposeLinter(
+            [EXTRA_JS_FILEPATH], FILE_CACHE)
+        lint_task_report = linter.check_extra_js_files()
+        self.assertEqual([
+            'core/templates/demo.js  --> Found extra .js file',
+            'If you want the above files to be present as js files, add '
+            'them to the list JS_FILEPATHS_NOT_TO_BUILD in build.py. '
+            'Otherwise, rename them to .ts'], lint_task_report.trimmed_messages)
+        self.assertEqual('Extra JS files', lint_task_report.name)
         self.assertTrue(lint_task_report.failed)
 
     def test_with_excluded_filepath(self):
